switch from CLEAR to PATCH

+++ relengapi/blueprints/tooltool/__init__.py
@@ -183,26 +183,37 @@ def get_file(algorithm, digest):
     return row.to_json()
 
 
-@bp.route('/file/<algorithm>/<digest>', methods=['CLEAR'])
+@bp.route('/file/<algorithm>/<digest>', methods=['PATCH'])
 @p.tooltool.manage.require()
-@api.apimethod(types.File, unicode, unicode)
-def clear_file(algorithm, digest):
-    """Clear all instances of the given file.  The file itself remains, but
-    until and unless someone uploads a new copy, it will not be available for
-    download."""
+@api.apimethod(types.File, unicode, unicode, body=[{unicode: unicode}])
+def patch_file(algorithm, digest, body):
+    """Make administrative changes to an existing file.  The body is a list of
+    changes to apply, each represented by a JSON object.
+
+    The object ``{"op": "delete_instances"}`` will cause all instances of the
+    file to be deleted.  The file record itself will not be deleted, as it is
+    still a part of one or more upload batches, but until and unless someone
+    uploads a new copy, the content will not be available for download."""
     session = current_app.db.session('tooltool')
     if algorithm != 'sha512':
         raise NotFound("Unknown algorithm")
     row = session.query(tables.File).filter(tables.File.sha512 == digest).first()
     if not row:
         raise NotFound
-    key_name = '/{}/{}'.format(algorithm, digest)
-    cfg = current_app.config.get('TOOLTOOL_REGIONS')
-    for i in row.instances:
-        conn = current_app.aws.connect_to('s3', i.region)
-        bucket = conn.get_bucket(cfg[i.region])
-        bucket.delete_key(key_name)
-        session.delete(i)
+
+    for change in body:
+        if 'op' not in change:
+            raise BadRequest("no op")
+        if change['op'] == 'delete_instances':
+            key_name = '/{}/{}'.format(algorithm, digest)
+            cfg = current_app.config.get('TOOLTOOL_REGIONS')
+            for i in row.instances:
+                conn = current_app.aws.connect_to('s3', i.region)
+                bucket = conn.get_bucket(cfg[i.region])
+                bucket.delete_key(key_name)
+                session.delete(i)
+        else:
+            raise BadRequest("unknown op")
     session.commit()
     # TODO: include instances here
     return row.to_json()
